Individual arch files and some edited scripts
For those who prefer the individual archetype databases instead of the bulk one, they are here for your convenience. (And probably mine too, so I can copy/paste codes as needed)

Also a few updates on old code in the TCG legacy support, and removal of some old images that are not needed.